+++ evtstrd/util.py
@@ -10,6 +10,7 @@ async def read_json_line(reader: StreamReader) -> Any:
     while True:
         line = await reader.readline()
         if line:
+            logging.debug(f"read line from socket: {line}")
             try:
                 return json.loads(line.decode("utf-8").strip())
             except (ValueError, UnicodeDecodeError):